Resolve issue #437 - Fix animation bug in [SignIn] (#438)
`transform` animation value is not preserved between render on mobile. This is the reason that `scale` is return to origin when writing a text on `textInput` which invokes `re-render`.  So I changed animation value from `scale` to `width` and `height`.

But there is a bug about `withSpring` function on `web`, then I splits an animation value according to platform. 
This bug seems to have related with `react-native-web`.  You can check relate issue with [here](https://github.com/software-mansion/react-native-reanimated/issues/1804) or [here](https://github.com/necolas/react-native-web/issues/1935).